Move tests to DecisionServiceTest.java. User Profile save happens in decision service.
