+++ core-api/src/test/java/com/optimizely/ab/OptimizelyTest.java
@@ -87,7 +87,6 @@ import static org.mockito.Matchers.eq;
 import static org.mockito.Mockito.doThrow;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.never;
-import static org.mockito.Mockito.spy;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
@@ -1981,35 +1980,6 @@ public class OptimizelyTest {
 
         assertNull(optimizely.getVariation(experiment.getKey(), "user",
                 Collections.singletonMap("browser_type", "firefox")));
-    }
-
-    /**
-     * Verify that {@link Optimizely#getVariation(String, String, Map)} gives precedence to forced variation bucketing
-     * over user profile and audience evaluation for murmurhash3 bucketing.
-     */
-    @Test
-    public void getVariationForcedVariationPrecedesUserProfileAndAudienceEval() throws Exception {
-        Bucketer bucketer = spy(new Bucketer(validProjectConfig));
-        Experiment experiment = validProjectConfig.getExperiments().get(0);
-        Variation expectedVariation = experiment.getVariations().get(0);
-        String whitelistedUserId = "testUser1";
-
-        Optimizely optimizely = Optimizely.builder(validDatafile, mockEventHandler)
-                .withBucketing(bucketer)
-                .withConfig(validProjectConfig)
-                .build();
-
-        // user excluded without audiences and whitelisting
-        assertNull(optimizely.getVariation(experiment.getKey(), genericUserId));
-
-        logbackVerifier.expectMessage(Level.INFO, "User \"" + whitelistedUserId + "\" is forced in variation \"vtag1\".");
-
-        // no attributes provided for a experiment that has an audience
-        assertThat(optimizely.getVariation(experiment.getKey(), whitelistedUserId), is(expectedVariation));
-
-        verify(bucketer).getForcedVariation(experiment, whitelistedUserId);
-        verify(bucketer, never()).getStoredVariation(experiment, whitelistedUserId);
-        verify(bucketer, never()).bucket(experiment, whitelistedUserId);
     }
 
     /**
